requirements.txt
flask
requests
python-dotenv
rich
numpy
